modify segment falt and implemet sharing-resource algorithms

+++ extern/simulte/src/stack/mac/layer/LteMacVUeMode4.h
@@ -115,7 +115,7 @@ protected:
    simsignal_t rriValue;
    bool isPassSPS;
    int selectedSubchannel_;
-   float selectedSlot_;
+   double selectedSlot_;
    std::vector<std::pair<int,simtime_t>> correspondingNode_;
 
 //   // Lte AMC module
